+++ frontend/public/kebab_logo.ico
