+++ frontend/src/App.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import AppRouter from './router'
+
+export default function App() {
+  return (
+    <AppRouter />
+  )
+}
