+++ frontend/src/img/kebab_logo.png
