--- kebab_city_tour/src/App.js
@@ -1,8 +1,0 @@
-import React from "react";
-import AppRouter from "./router";
-
-export default function App() {
-  return (
-    <AppRouter />
-  );
-}
