--- kebab_city_tour/src/router.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Home from './pages/Home';
-import Map from './pages/Map';
-import LoginPage from './pages/LoginPage';
-
-const AppRouter = () => (
-  <Router>
-    <Routes>
-      <Route path="/" element={<Home />} />
-      <Route path="/map" element={<Map />} />
-      <Route path="/login" element={<LoginPage />} />
-    </Routes>
-  </Router>
-);
-
-export default AppRouter;
